requirements.txt
pyTelegramBotAPI
requests
beautifulsoup4
Pillow
reportlab
flask
